+++ apps/bfd-server/bfd-server-war/src/main/java/gov/cms/bfd/server/war/stu3/providers/TransformerUtils.java
@@ -863,7 +863,7 @@ public final class TransformerUtils {
       throw new NoSuchElementException();
     }
     try {
-      String stringDate = String.format("%0" + 4 + "d", dateYear.get());
+      String stringDate = String.format("%0" + 4 + "d", dateYear.get().intValue());
       ;
       DateType dateYearValue = new DateType(stringDate);
       String extensionUrl = calculateVariableReferenceUrl(ccwVariable);
